fix: Fixing reflow pipeline errors

+++ inc/logger.h
@@ -1,7 +1,9 @@
 #ifndef LOG_H_
 #define LOG_H_
 
-void escreve_temp_log(int TI, int TR, int TE);
+#include<string>
+
+void escreve_temp_log(float TI, float TR, float TE);
 void escreve_string_log(std::string mensagem);
 void fecha_log();
 void inicia_log();
